styling/fix: fixed styling in some components
