+++ frontend/components/UI/BgOverlay.module.css
@@ -1,5 +1,6 @@
 .bg_overlay {
   position: fixed;
+  top: 100px;
   height: 100%;
   width: 100vw;
   background: rgba(41, 45, 50, 0.6);
