+++ frontend/components/layout/Navbar/Navbar.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import Link from "next/link";
 import Image from "next/image";
 import Logo from "../../assets/WorkshopAppLogo.svg";
 import HamburgerIcon from "../../assets/HamburgerIcon.svg";
@@ -18,11 +19,16 @@ const Navbar = () => {
   return (
     <>
       <header className={styles.header}>
-        <Image src={Logo} alt="logo" />
+        <Link href="/">
+          <Image src={Logo} alt="logo" />
+        </Link>
 
-        {/* <div onClick={handleNav}> */}
-        <Image src={HamburgerIcon} alt="hamburger" onClick={handleNav} />
-        {/* </div> */}
+        <Image
+          src={HamburgerIcon}
+          alt="hamburger"
+          onClick={handleNav}
+          className={styles.hamburger}
+        />
       </header>
       {isOpen && <NavList handleNav={handleNav} />}
       {isOpen && <BgOverlay handleBg={handleNav} />}
