+++ frontend/components/layout/Navbar/Navbar.module.css
@@ -1,8 +1,17 @@
 .header {
+  position: fixed;
+  top: 0;
+  width: 100vw;
+  background-color: white;
   display: flex;
   justify-content: center;
   align-items: flex-end;
   gap: 5px;
-  padding: 25px 0px;
+  padding: 30px 0px;
   box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.07);
+  z-index: 20;
+}
+
+.hamburger {
+  cursor: pointer;
 }
