+++ frontend/styles/About.module.css
@@ -36,7 +36,7 @@
 }
 
 .quote__icon__container2 {
-  width: 45%;
+  width: 20%;
   display: flex;
   align-items: flex-end;
   justify-content: flex-end;
